requirements.txt
# Core 3D processing and generation
trimesh==4.0.5          # 3D mesh generation, loading, and processing
numpy==1.24.3           # Numerical operations and array handling
open3d==0.18.0          # Advanced 3D processing and visualization

# API and web framework
fastapi==0.104.1        # REST API framework
uvicorn==0.24.0          # ASGI server for FastAPI
pydantic==2.5.0         # Data validation and serialization

# HTTP and networking
requests==2.31.0        # HTTP client for API calls
httpx==0.25.2           # Async HTTP client for testing

# Image processing
pillow==10.1.0          # Image processing for screenshots

# Testing framework
pytest==7.4.3           # Testing framework
pytest-asyncio==0.21.1  # Async testing support

# Utilities
tqdm==4.66.1            # Progress bars for long operations
